joule_rt::stream

Trait Stream

pub trait Stream {
    type Item;

    // Required method
    fn poll_next(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>,
    ) -> Poll<Option<Self::Item>>;

    // Provided method
    fn size_hint(&self) -> (usize, Option<usize>) { ... }
}

A stream of values produced asynchronously.

This is the async equivalent of Iterator.

Required Associated Types

type Item

The type of items yielded by the stream.

Required Methods

fn poll_next( self: Pin<&mut Self>, cx: &mut Context<'_>, ) -> Poll<Option<Self::Item>>

Attempt to pull out the next value of this stream.

Returns Poll::Ready(Some(item)) if an item is ready, Poll::Ready(None) if the stream is exhausted, or Poll::Pending if the next item is not yet available.

Provided Methods

fn size_hint(&self) -> (usize, Option<usize>)

Returns the bounds on the remaining length of the stream.